+++ core/src/main/java/org/opensearch/sql/calcite/utils/UserDefineFunctionUtils.java
@@ -82,7 +82,7 @@ public class UserDefineFunctionUtils {
   }
 
   public static SqlReturnTypeInference getReturnTypeInference(int targetPosition) {
-    SqlReturnTypeInference a = opBinding -> {
+    return opBinding -> {
       RelDataTypeFactory typeFactory = opBinding.getTypeFactory();
 
       // Get argument types
@@ -92,9 +92,7 @@ public class UserDefineFunctionUtils {
         throw new IllegalArgumentException("Function requires at least one argument.");
       }
       RelDataType firstArgType = argTypes.get(targetPosition);
-      return typeFactory.createSqlType(firstArgType.getSqlTypeName());
+      return typeFactory.createTypeWithNullability(typeFactory.createSqlType(firstArgType.getSqlTypeName()), true);
     };
-    a.andThen(SqlTypeTransforms.FORCE_NULLABLE);
-    return a;
   }
 }
